fix: SSE replay flood, Server Action errors in production, empty credentials in edit form
- SSE: only replay cached events on reconnect (Last-Event-ID), not on fresh page loads
- Server Actions: replace throw with safe return objects to avoid Next.js production error masking
- Integrations: always return decrypted credentials in GET /integrations/{id} for all users
- Invoicing: add order_number column, fix pagination, add reactive SSE insert with animation
- Table: add rowClassName prop for per-row styling

+++ back/central/services/events/internal/infra/secondary/sse/sse_message.go
@@ -49,9 +49,9 @@ func (m *EventManager) toString(v interface{}) string {
 	case int32:
 		return fmt.Sprintf("%d", val)
 	case float64:
-		return fmt.Sprintf("%.2f", val)
+		return fmt.Sprintf("%d", int64(val))
 	case float32:
-		return fmt.Sprintf("%.2f", val)
+		return fmt.Sprintf("%d", int64(val))
 	case bool:
 		if val {
 			return "true"
